refactor: split VideoPlayer's media layers into hooks (Q10)
Signed-URL minting/refresh/recovery is useSignedPlayback, subtitle
fetch and cue selection are useSubtitleTrack, page-crossing prev/next
is usePlaylistNavigation, and the pure VTT parser lands in
lib/subtitles.js. Deletes the long-unused useSignedUrlRefresh hook.
1,304 lines down to 891.
